test(http): cover nested uploaded files through the mapper
